+++ statistics-impl/src/main/java/com/dwarfeng/statistics/impl/handler/ResetProcessor.java
@@ -1,11 +1,10 @@
 package com.dwarfeng.statistics.impl.handler;
 
-import com.dwarfeng.statistics.stack.handler.DriveLocalCacheHandler;
-import com.dwarfeng.statistics.stack.handler.ExecuteLocalCacheHandler;
-import com.dwarfeng.statistics.stack.handler.ReceiveHandler;
-import com.dwarfeng.statistics.stack.handler.SuperviseHandler;
+import com.dwarfeng.statistics.stack.handler.*;
 import com.dwarfeng.subgrade.sdk.exception.HandlerExceptionHelper;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import java.util.concurrent.locks.Lock;
@@ -20,11 +19,15 @@ import java.util.concurrent.locks.ReentrantLock;
 @Component
 public class ResetProcessor {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(ResetProcessor.class);
+
     private final SuperviseHandler superviseHandler;
     private final DriveLocalCacheHandler driveLocalCacheHandler;
 
     private final ReceiveHandler receiveHandler;
     private final ExecuteLocalCacheHandler executeLocalCacheHandler;
+
+    private final PushHandler pushHandler;
 
     private final Lock lock = new ReentrantLock();
 
@@ -32,12 +35,14 @@ public class ResetProcessor {
             SuperviseHandler superviseHandler,
             DriveLocalCacheHandler driveLocalCacheHandler,
             ReceiveHandler receiveHandler,
-            ExecuteLocalCacheHandler executeLocalCacheHandler
+            ExecuteLocalCacheHandler executeLocalCacheHandler,
+            PushHandler pushHandler
     ) {
         this.superviseHandler = superviseHandler;
         this.driveLocalCacheHandler = driveLocalCacheHandler;
         this.receiveHandler = receiveHandler;
         this.executeLocalCacheHandler = executeLocalCacheHandler;
+        this.pushHandler = pushHandler;
     }
 
     public void resetSupervise() throws HandlerException {
@@ -63,6 +68,13 @@ public class ResetProcessor {
         if (superviseHandlerStarted) {
             superviseHandler.start();
         }
+
+        // 消息推送。
+        try {
+            pushHandler.superviseReset();
+        } catch (Exception e) {
+            LOGGER.warn("推送主管功能重置消息时发生异常, 本次消息将不会被推送, 异常信息如下: ", e);
+        }
     }
 
     public void resetExecute() throws HandlerException {
@@ -87,6 +99,13 @@ public class ResetProcessor {
         // 如果接收处理器之前是启动的，则重新启动。
         if (receiveHandlerStarted) {
             receiveHandler.start();
+        }
+
+        // 消息推送。
+        try {
+            pushHandler.executeReset();
+        } catch (Exception e) {
+            LOGGER.warn("推送执行功能重置消息时发生异常, 本次消息将不会被推送, 异常信息如下: ", e);
         }
     }
 }